refactor(foundation): introduce XMLFileDeserializer
