+++ components/ILIAS/Test/classes/class.ilTestImporter.php
@@ -21,7 +21,7 @@ declare(strict_types=1);
 use ILIAS\Data\ReferenceId;
 use ILIAS\Test\ExportImport\Import\TestImporter;
 use ILIAS\TestQuestionPool\ExportImport\Foundation\Importing\ImportSessionRepository;
-use ILIAS\TestQuestionPool\ExportImport\Foundation\Serializing\SimpleXMLDeserializer;
+use ILIAS\TestQuestionPool\ExportImport\Foundation\Serializing\XMLMemoryDeserializer;
 use ILIAS\TestQuestionPool\ExportImport\Import\DetectLegacyImportStage;
 use ILIAS\Test\TestDIC;
 
@@ -70,7 +70,7 @@ class ilTestImporter extends ilXmlImporter
         }
 
         $result = $this->importer->import(
-            new SimpleXMLDeserializer()->open($a_xml),
+            new XMLMemoryDeserializer()->open($a_xml),
             $a_mapping,
             new ReferenceId($a_mapping->getTargetId()),
             $context,
