+++ components/ILIAS/Test/src/ExportImport/Import/PersistStage.php
@@ -25,7 +25,7 @@ use ILIAS\TestQuestionPool\ExportImport\Foundation\Contracts\ImportStage;
 use ILIAS\TestQuestionPool\ExportImport\Foundation\Importing\ImportContext;
 use ILIAS\TestQuestionPool\ExportImport\Foundation\Importing\ImportSessionRepository;
 use ILIAS\TestQuestionPool\ExportImport\Foundation\Importing\StageResult;
-use ILIAS\TestQuestionPool\ExportImport\Foundation\Serializing\SimpleXMLDeserializer;
+use ILIAS\TestQuestionPool\ExportImport\Foundation\Serializing\XMLFileDeserializer;
 use ILIAS\TestQuestionPool\ExportImport\Import\UploadValidationStage;
 use ilImport;
 
@@ -66,7 +66,7 @@ class PersistStage implements ImportStage
             return StageResult::error($context, $this->lng->txt('obj_import_file_error'));
         }
 
-        $deserializer = new SimpleXMLDeserializer()->open(file_get_contents($mappings_file));
+        $deserializer = new XMLFileDeserializer()->open($mappings_file);
         $deserializer->addHandler('mappings', function (array $mappings) use (&$context) {
             $context = $context->with('mappings', $mappings);
         });
